+++ .vscode/settings.json
@@ -1,0 +1,6 @@
+{
+    "files.associations": {
+        "sdl_render.h": "c",
+        "linkedlist.h": "c"
+    }
+}
