+++ src/include/grid.h
@@ -1,0 +1,8 @@
+#include "graphics.h"
+
+#ifndef GRID_H
+#define GRID_H
+
+void renderGrid();
+
+#endif
