+++ src/include/position.h
@@ -1,0 +1,10 @@
+
+#ifndef POSITION_H
+#define POSITION_H
+
+typedef struct {
+    int x;
+    int y;
+} Position;
+
+#endif
